Refactor entry point structure
- Moved the main application logic to app/__main__.py and use main.py to serve as a convenience wrapper.
- Add a pyproject.toml to be able to start the program with "topoptcomec" instead of "python main.py"
- Take care or documentation

Advantages & Impacts:
Provides a standard, package-installable entry point and isolates dependencies within the virtual environment without affecting start times or performance.
