+++ .vscode/launch.json
@@ -5,7 +5,7 @@
             "name": "Python: Run Main",
             "type": "python",
             "request": "launch",
-            "program": "${workspaceFolder}/main.py",
+            "module": "app",
             "console": "integratedTerminal",
             "cwd": "${workspaceFolder}",
             "justMyCode": true
