+++ pyproject.toml
@@ -1,0 +1,51 @@
+[build-system]
+requires = ["setuptools>=68.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "topoptcomec"
+version = "0.1.0"
+description = "Interactive topology optimization for compliant mechanisms"
+readme = "README.md"
+license = {text = "MIT"}
+authors = [{name = "Luc Prevost"}]
+keywords = ["topology-optimization", "compliant-mechanisms", "engineering"]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
+]
+requires-python = ">=3.10"
+dependencies = [
+    "pyside6",
+    "numpy",
+    "scipy",
+    "matplotlib",
+    "vtk",
+    "PyMCubes",
+    "numpy-stl",
+    "lib3mf",
+]
+
+[project.urls]
+Homepage = "https://github.com/ninja7V/topoptcomec"
+Repository = "https://github.com/ninja7V/topoptcomec.git"
+Documentation = "https://github.com/ninja7v/TopoptComec/wiki"
+Issues = "https://github.com/ninja7V/topoptcomec/issues"
+
+[project.scripts]
+topoptcomec = "app.__main__:main"
+
+[tool.setuptools]
+packages = ["app", "app.core", "app.ui"]
+include-package-data = true
+
+[tool.setuptools.package-data]
+app = ["icons/*"]
